Movex ExaBGP into separate module
