+++ src/bgp_protocol.cpp
@@ -27,11 +27,6 @@ std::ostream& operator<<(std::ostream& os, bgp_flow_spec_protocol_t const& proto
     }
 }
 
-void exabgp_flow_spec_rule_ban_manage(std::string action, flow_spec_rule_t flow_spec_rule) {
-    // "announce flow route {\\n match {\\n source 10.0.0.1/32;\\nsource-port =" + str(i) +
-    // ";\\n destination 1.2.3.4/32;\\n }\\n then {\\n discard;\\n }\\n }\\n\n")
-}
-
 std::ostream& operator<<(std::ostream& os, flow_spec_tcp_flags_t const& tcp_flag) {
     if (tcp_flag == FLOW_TCP_FLAG_SYN) {
         return os << "syn";
