+++ src/fastnetmon_logic.cpp
@@ -32,6 +32,8 @@
 #include "actions/gobgp_action.h"
 #endif
 
+#include "actions/exabgp_action.hpp"
+
 // Yes, maybe it's not an good idea but with this we can guarantee working code in example plugin
 #include "example_plugin/example_collector.h"
 
@@ -142,12 +144,6 @@ extern map_for_subnet_counters_t PerSubnetAverageSpeedMap;
 extern bool enable_subnet_counters;
 extern ban_settings_t global_ban_settings;
 extern bool exabgp_enabled;
-extern std::string exabgp_community;
-extern std::string exabgp_community_subnet;
-extern std::string exabgp_community_host;
-extern std::string exabgp_command_pipe;
-extern std::string exabgp_next_hop;
-extern bool exabgp_announce_host;
 extern bool exabgp_flow_spec_announces;
 extern bool gobgp_enabled;
 extern map_of_vector_counters_t SubnetVectorMapSpeedAverage;
@@ -1334,7 +1330,7 @@ void launch_bgp_flow_spec_rule(amplification_attack_type_t attack_type, std::str
             }
         } else {
             // We have already blocked this attack
-            logger << log4cpp::Priority::INFO << "The same rule was already sent to ExaBGP formerly";
+            logger << log4cpp::Priority::INFO << "The same rule was already sent to ExaBGP previously";
         }
     } else {
         logger << log4cpp::Priority::INFO << "exabgp_flow_spec_announces disabled. We will not talk to ExaBGP";
@@ -1675,91 +1671,6 @@ ban_settings_t get_ban_settings_for_this_subnet(subnet_cidr_mask_t subnet, std::
     // We found ban settings for this host group and use they instead global
     return hostgroup_settings_itr->second;
 }
-
-
-void exabgp_ban_manage(std::string action, std::string ip_as_string, attack_details_t current_attack) {
-    // We will announce whole subent here
-    if (exabgp_announce_whole_subnet) {
-        std::string subnet_as_string_with_mask = convert_subnet_to_string(current_attack.customer_network);
-
-        exabgp_prefix_ban_manage(action, subnet_as_string_with_mask, exabgp_next_hop, exabgp_community_subnet);
-    }
-
-    // And we could announce single host here (/32)
-    if (exabgp_announce_host) {
-        std::string ip_as_string_with_mask = ip_as_string + "/32";
-
-        exabgp_prefix_ban_manage(action, ip_as_string_with_mask, exabgp_next_hop, exabgp_community_host);
-    }
-}
-
-
-// Low level ExaBGP ban management
-void exabgp_prefix_ban_manage(std::string action,
-                              std::string prefix_as_string_with_mask,
-                              std::string exabgp_next_hop,
-                              std::string exabgp_community) {
-
-    /* Buffer for BGP message */
-    char bgp_message[256];
-
-    if (action == "ban") {
-        sprintf(bgp_message, "announce route %s next-hop %s community %s\n",
-                prefix_as_string_with_mask.c_str(), exabgp_next_hop.c_str(), exabgp_community.c_str());
-    } else {
-        sprintf(bgp_message, "withdraw route %s next-hop %s\n", prefix_as_string_with_mask.c_str(),
-                exabgp_next_hop.c_str());
-    }
-
-    logger << log4cpp::Priority::INFO << "ExaBGP announce message: " << bgp_message;
-
-    int exabgp_pipe = open(exabgp_command_pipe.c_str(), O_WRONLY);
-
-    if (exabgp_pipe <= 0) {
-        logger << log4cpp::Priority::ERROR << "Can't open ExaBGP pipe " << exabgp_command_pipe
-               << " Ban is not executed";
-        return;
-    }
-
-    int wrote_bytes = write(exabgp_pipe, bgp_message, strlen(bgp_message));
-
-    if (wrote_bytes != strlen(bgp_message)) {
-        logger << log4cpp::Priority::ERROR << "Can't write message to ExaBGP pipe";
-    }
-
-    close(exabgp_pipe);
-}
-
-bool exabgp_flow_spec_ban_manage(std::string action, std::string flow_spec_rule_as_text) {
-    std::string announce_action;
-
-    if (action == "ban") {
-        announce_action = "announce";
-    } else {
-        announce_action = "withdraw";
-    }    
-
-    // Trailing \n is very important!
-    std::string bgp_message = announce_action + " " + flow_spec_rule_as_text + "\n";
-
-    int exabgp_pipe = open(exabgp_command_pipe.c_str(), O_WRONLY);
-
-    if (exabgp_pipe <= 0) { 
-        logger << log4cpp::Priority::ERROR << "Can't open ExaBGP pipe for flow spec announce " << exabgp_command_pipe;
-        return false;
-    }    
-
-    int wrote_bytes = write(exabgp_pipe, bgp_message.c_str(), bgp_message.size());
-
-    if (wrote_bytes != bgp_message.size()) {
-        logger << log4cpp::Priority::ERROR << "Can't write message to ExaBGP pipe";
-        return false;
-    }    
-
-    close(exabgp_pipe);
-    return true;
-}
-
 
 #ifdef REDIS
 void store_data_in_redis(std::string key_name, std::string attack_details) {
