+++ src/fastnetmon_logic.hpp
@@ -88,7 +88,6 @@ uint64_t convert_conntrack_hash_struct_to_integer(packed_conntrack_hash_t* struc
 bool process_flow_tracking_table(conntrack_main_struct_t& conntrack_element, std::string client_ip);
 bool exec_with_stdin_params(std::string cmd, std::string params);
 ban_settings_t get_ban_settings_for_this_subnet(subnet_cidr_mask_t subnet, std::string& host_group_name);
-void exabgp_ban_manage(std::string action, std::string ip_as_string, attack_details_t current_attack);
 void exabgp_prefix_ban_manage(std::string action,
                               std::string prefix_as_string_with_mask,
                               std::string exabgp_next_hop,
